Use 'TOTMEM' to increase the available memory with dynamic allocation
The current memory allocation scheme uses 'MEMSIZ' to
determine the memory allocation size, it is not the intended
use of 'MEMSIZ'.  The real purpose of 'MEMSIZ' is to hold
the count of total number of array/matrix elements that
can be used in a particular process without overwriting
the 'TOTAL MEMORY' available to it.

Use 'TOTMEM' to increase the memory limit in the case of
memory allocation with DYNAMIC_WORK_MEM_ALLOC instead
of 'MEMSIZ'

Also add the missing 'free()' in few places.

Change-Id: I23d7549fcb288de2aa540c63b3d4a401c5e58531