PDSYEVD Optimization - NwChem
The NwChem application reported a degradation in performance with 192
ranks. On further investigation, it was found that the application used
only 191 processes for ScaLAPACK with grid dimensions (1 x 191) which
was a flat grid configuration.

Flat grids performed poorly within the PDYSTRD API with more time spent
on communication. An alternative to this reduction, PDSYNTRD was opted.

Hence, The PDSYEVD routine was optimized for flat/rectangular grids with
the use of PDSYNTRD API for tridiagonal reduction.

The configuration for Nwchem has been tested on standalone code and the
test cases are upadted in TESTING/EXT_TEST/SEP.dat
Since moderate sizes are being tested, we need to build ScaLAPACK with
'-DENABLE_LARGE_MATRIX_TESTING=ON'

Signed-off-by: Sundari H <Sundari.SwathyMeenaH@amd.com>
AMD-Internal: [CPUPL-5087]

Change-Id: I972fa333560f0c6507649d8794adc32c54340f6b